# Rewrite prettyText to fix performance issue.