Following this [discussion on reddit](https://redd.it/3ye4o0) it appears that `prettyText` is really slow and/or as a space leak. The actual go using `Text.drop` repeatedly, which according to the doc is O(n) which should result in the whole to be O(n^2) which will explain the poor performance.

This patch fix the performance issue by using `foldr` which should be in O(n) and constant space (as manual benchmark shows). I added a simple test program on this [branch](https://github.com/maxigit/lucid-svg/tree/bench-pretty). On my computer

  + bench p    10000 (normal pretty) 132s
  + bench       10000 (normal non-pretty) take about 0.17s
  + bench p    10000 (new version) about 1.06 s
  + bench    100,000 (normal non-pretty) 1.27 
  + bench p 100,000 (new version) 10.71s

(to test the new version, you'll need to merge this branch with the `bench-pretty` branch.